+++ enclone_visual/src/process_messages.rs
@@ -104,35 +104,35 @@ impl EncloneVisual {
                         let hi = self.h.history_index;
                         self.h
                             .input1_history
-                            .insert(hi, self.h.input1_hist_uniq.len() as u32);
+                            .insert(hi as usize, self.h.input1_hist_uniq.len() as u32);
                         self.h.input1_hist_uniq.push(self.input1_value.clone());
                         self.h
                             .input2_history
-                            .insert(hi, self.h.input2_hist_uniq.len() as u32);
+                            .insert(hi as usize, self.h.input2_hist_uniq.len() as u32);
                         self.h.input2_hist_uniq.push(self.input2_value.clone());
                         self.h
                             .translated_input_history
-                            .insert(hi, self.h.translated_input_hist_uniq.len() as u32);
+                            .insert(hi as usize, self.h.translated_input_hist_uniq.len() as u32);
                         self.h
                             .translated_input_hist_uniq
                             .push(args2.iter().format(" ").to_string());
-                        self.h.svg_history.insert(hi, self.h.svg_history[hi - 1]);
+                        self.h.svg_history.insert(hi as usize, self.h.svg_history[(hi - 1) as usize]);
                         self.h
                             .summary_history
-                            .insert(hi, self.h.summary_history[hi - 1]);
+                            .insert(hi as usize, self.h.summary_history[(hi - 1) as usize]);
                         self.h
                             .displayed_tables_history
-                            .insert(hi, self.h.displayed_tables_hist_uniq.len() as u32);
+                            .insert(hi as usize, self.h.displayed_tables_hist_uniq.len() as u32);
                         self.h
                             .displayed_tables_hist_uniq
                             .push(reply_text.to_string());
                         self.h
                             .table_comp_history
-                            .insert(hi, self.h.table_comp_history[hi - 1]);
+                            .insert(hi as usize, self.h.table_comp_history[(hi - 1) as usize]);
                         self.h
                             .last_widths_history
-                            .insert(hi, self.h.last_widths_history[hi - 1]);
-                        self.h.is_blank.insert(hi, self.is_blank_current());
+                            .insert(hi as usize, self.h.last_widths_history[(hi - 1) as usize]);
+                        self.h.is_blank.insert(hi as usize, self.is_blank_current());
                         self.h.history_index += 1;
                         if !TEST_MODE.load(SeqCst) {
                             Command::none()
@@ -195,15 +195,15 @@ impl EncloneVisual {
 
             Message::DelButtonPressed(_) => {
                 let h = self.h.history_index - 1;
-                self.h.svg_history.remove(h);
-                self.h.summary_history.remove(h);
-                self.h.input1_history.remove(h);
-                self.h.input2_history.remove(h);
-                self.h.translated_input_history.remove(h);
-                self.h.displayed_tables_history.remove(h);
-                self.h.table_comp_history.remove(h);
-                self.h.last_widths_history.remove(h);
-                self.h.is_blank.remove(h);
+                self.h.svg_history.remove(h as usize);
+                self.h.summary_history.remove(h as usize);
+                self.h.input1_history.remove(h as usize);
+                self.h.input2_history.remove(h as usize);
+                self.h.translated_input_history.remove(h as usize);
+                self.h.displayed_tables_history.remove(h as usize);
+                self.h.table_comp_history.remove(h as usize);
+                self.h.last_widths_history.remove(h as usize);
+                self.h.is_blank.remove(h as usize);
                 if self.state_count() == 0 {
                     self.h.history_index -= 1;
                     self.input1_value.clear();
@@ -411,9 +411,9 @@ impl EncloneVisual {
                 let hi = self.h.history_index;
                 let len = self.h.table_comp_hist_uniq.len();
                 if len > 0 && self.h.table_comp_hist_uniq[len - 1] == reply_table_comp {
-                    self.h.table_comp_history.insert(hi, (len - 1) as u32);
+                    self.h.table_comp_history.insert(hi as usize, (len - 1) as u32);
                 } else {
-                    self.h.table_comp_history.insert(hi, len as u32);
+                    self.h.table_comp_history.insert(hi as usize, len as u32);
                     self.h.table_comp_hist_uniq.push(reply_table_comp.clone());
                     if self.table_comp_value.len() > 0 {
                         let mut gunzipped = Vec::<u8>::new();
@@ -450,58 +450,58 @@ impl EncloneVisual {
 
                 let len = self.h.last_widths_hist_uniq.len();
                 if len > 0 && self.h.last_widths_hist_uniq[len - 1] == reply_last_widths {
-                    self.h.last_widths_history.insert(hi, (len - 1) as u32);
+                    self.h.last_widths_history.insert(hi as usize, (len - 1) as u32);
                 } else {
-                    self.h.last_widths_history.insert(hi, len as u32);
+                    self.h.last_widths_history.insert(hi as usize, len as u32);
                     self.h.last_widths_hist_uniq.push(reply_last_widths.clone());
                 }
                 let len = self.h.svg_hist_uniq.len();
                 if len > 0 && self.h.svg_hist_uniq[len - 1] == reply_svg {
-                    self.h.svg_history.insert(hi, (len - 1) as u32);
+                    self.h.svg_history.insert(hi as usize, (len - 1) as u32);
                 } else {
-                    self.h.svg_history.insert(hi, len as u32);
+                    self.h.svg_history.insert(hi as usize, len as u32);
                     self.h.svg_hist_uniq.push(reply_svg.clone());
                 }
                 let len = self.h.summary_hist_uniq.len();
                 if len > 0 && self.h.summary_hist_uniq[len - 1] == reply_summary {
-                    self.h.summary_history.insert(hi, (len - 1) as u32);
+                    self.h.summary_history.insert(hi as usize, (len - 1) as u32);
                 } else {
-                    self.h.summary_history.insert(hi, len as u32);
+                    self.h.summary_history.insert(hi as usize, len as u32);
                     self.h.summary_hist_uniq.push(reply_summary.clone());
                 }
                 let len = self.h.displayed_tables_hist_uniq.len();
                 if len > 0 && self.h.displayed_tables_hist_uniq[len - 1] == reply_text {
-                    self.h.displayed_tables_history.insert(hi, (len - 1) as u32);
+                    self.h.displayed_tables_history.insert(hi as usize, (len - 1) as u32);
                 } else {
-                    self.h.displayed_tables_history.insert(hi, len as u32);
+                    self.h.displayed_tables_history.insert(hi as usize, len as u32);
                     self.h.displayed_tables_hist_uniq.push(reply_text.clone());
                 }
                 let len = self.h.input1_hist_uniq.len();
                 if len > 0 && self.h.input1_hist_uniq[len - 1] == self.input1_value {
-                    self.h.input1_history.insert(hi, (len - 1) as u32);
+                    self.h.input1_history.insert(hi as usize, (len - 1) as u32);
                 } else {
-                    self.h.input1_history.insert(hi, len as u32);
+                    self.h.input1_history.insert(hi as usize, len as u32);
                     self.h.input1_hist_uniq.push(self.input1_value.clone());
                 }
                 let len = self.h.input2_hist_uniq.len();
                 if len > 0 && self.h.input2_hist_uniq[len - 1] == self.input2_value {
-                    self.h.input2_history.insert(hi, (len - 1) as u32);
+                    self.h.input2_history.insert(hi as usize, (len - 1) as u32);
                 } else {
-                    self.h.input2_history.insert(hi, len as u32);
+                    self.h.input2_history.insert(hi as usize, len as u32);
                     self.h.input2_hist_uniq.push(self.input2_value.clone());
                 }
                 let len = self.h.translated_input_hist_uniq.len();
                 if len > 0
                     && self.h.translated_input_hist_uniq[len - 1] == self.translated_input_value
                 {
-                    self.h.translated_input_history.insert(hi, (len - 1) as u32);
+                    self.h.translated_input_history.insert(hi as usize, (len - 1) as u32);
                 } else {
-                    self.h.translated_input_history.insert(hi, len as u32);
+                    self.h.translated_input_history.insert(hi as usize, len as u32);
                     self.h
                         .translated_input_hist_uniq
                         .push(self.translated_input_value.clone());
                 }
-                self.h.is_blank.insert(hi, blank);
+                self.h.is_blank.insert(hi as usize, blank);
                 self.h.history_index += 1;
                 self.output_value = reply_text.to_string();
                 self.svg_value = reply_svg.to_string();
@@ -560,7 +560,7 @@ impl EncloneVisual {
                 if verbose {
                     xprintln!("\ncapturing, input history:");
                     for i in 0..self.h.input1_history.len() {
-                        let mark = if i + 1 == self.h.history_index {
+                        let mark = if i + 1 == self.h.history_index as usize {
                             "*"
                         } else {
                             ""
